Move stuff around to prepare for tests
